+++ src/main/java/com/github/t45k/abcd/ast/constructor/ASTConstructor.java
@@ -1,18 +1,74 @@
 package com.github.t45k.abcd.ast.constructor;
 
-import org.eclipse.jdt.core.dom.AST;
+import com.github.t45k.abcd.ast.FileAST;
+import org.eclipse.core.runtime.NullProgressMonitor;
+import org.eclipse.jdt.core.JavaCore;
 import org.eclipse.jdt.core.dom.ASTParser;
+import org.eclipse.jdt.core.dom.CompilationUnit;
+import org.eclipse.jdt.core.dom.FileASTRequestor;
+import org.eclipse.jdt.core.formatter.DefaultCodeFormatterConstants;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
-public class ASTConstructor extends AbstractASTConstructor {
+public abstract class ASTConstructor implements IASTConstructor {
+    static final String INVALID_PATH_EXCEPTION_MESSAGE = "Invalid path was specified";
+
+    public static ASTConstructor create(final Path targetFilesRootPath, final Path classpathRootPath, final Path libRootPath) {
+        if (classpathRootPath == null || libRootPath == null) {
+            return new NonBindingASTConstructor();
+        }
+        return new BindingASTConstructor(targetFilesRootPath, classpathRootPath, libRootPath);
+    }
 
     @Override
-    public ASTParser createParser() {
-        final ASTParser parser = ASTParser.newParser(AST.JLS10);
-        final Map<String, String> options = getOptionMap();
-        parser.setCompilerOptions(options);
+    public List<FileAST> constructFileAST(final Path targetFilesRootPath) {
+        final ASTParser parser = createParser();
 
-        return parser;
+        final List<FileAST> fileASTList = new ArrayList<>();
+        final FileASTRequestor requestor = new FileASTRequestor() {
+            @Override
+            public void acceptAST(final String sourceFilePath, final CompilationUnit ast) {
+                ast.recordModifications();
+                fileASTList.add(new FileAST(Paths.get(sourceFilePath), ast));
+            }
+        };
+
+        final List<Path> targetFilePathList = getTargetFiles(targetFilesRootPath, ".java");
+        final String[] sourceFilePaths = convertPathListToStringArray(targetFilePathList);
+        parser.createASTs(sourceFilePaths, null, new String[]{}, requestor, new NullProgressMonitor());
+
+        return fileASTList;
+    }
+
+    @SuppressWarnings("unchecked")
+    Map<String, String> getOptionMap() {
+        final Map<String, String> options = DefaultCodeFormatterConstants.getEclipseDefaultSettings();
+        options.put(JavaCore.COMPILER_COMPLIANCE, JavaCore.VERSION_10);
+        options.put(JavaCore.COMPILER_CODEGEN_TARGET_PLATFORM, JavaCore.VERSION_10);
+        options.put(JavaCore.COMPILER_SOURCE, JavaCore.VERSION_10);
+        return options;
+    }
+
+    List<Path> getTargetFiles(final Path targetFileRootPath, final String targetExtension) {
+        try {
+            return Files.walk(targetFileRootPath)
+                    .filter(path -> path.toString().endsWith(targetExtension))
+                    .collect(Collectors.toList());
+        } catch (IOException e) {
+            throw new RuntimeException(INVALID_PATH_EXCEPTION_MESSAGE);
+        }
+    }
+
+    String[] convertPathListToStringArray(final List<Path> pathList) {
+        return pathList.stream()
+                .map(Path::toString)
+                .toArray(String[]::new);
     }
 }
