--- src/main/java/com/github/t45k/abcd/ast/constructor/AbstractASTConstructor.java
@@ -1,78 +1,0 @@
-package com.github.t45k.abcd.ast.constructor;
-
-import com.github.t45k.abcd.ast.FileAST;
-import org.eclipse.core.runtime.NullProgressMonitor;
-import org.eclipse.jdt.core.JavaCore;
-import org.eclipse.jdt.core.dom.ASTParser;
-import org.eclipse.jdt.core.dom.CompilationUnit;
-import org.eclipse.jdt.core.dom.FileASTRequestor;
-import org.eclipse.jdt.core.formatter.DefaultCodeFormatterConstants;
-
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-public abstract class AbstractASTConstructor implements IASTConstructor{
-    static final String INVALID_PATH_EXCEPTION_MESSAGE = "Invalid path was specified";
-
-    public static AbstractASTConstructor create(final Path targetFilesRootPath, final Path classpathRootPath, final Path libRootPath) {
-        if (classpathRootPath == null || libRootPath == null) {
-            return new ASTConstructor();
-        } else {
-            return new BindingASTConstructor(targetFilesRootPath, classpathRootPath, libRootPath);
-        }
-    }
-
-    @Override
-    public List<FileAST> constructFileAST(final Path targetFilesRootPath) {
-        final ASTParser parser = createParser();
-
-        final List<FileAST> fileASTList = new ArrayList<>();
-        final FileASTRequestor requestor = new FileASTRequestor() {
-            @Override
-            public void acceptAST(final String sourceFilePath, final CompilationUnit ast) {
-                ast.recordModifications();
-                fileASTList.add(new FileAST(Paths.get(sourceFilePath), ast));
-            }
-        };
-
-        final List<Path> targetFilePathList = getTargetFiles(targetFilesRootPath, ".java");
-        final String[] sourceFilePaths = convertPathListToStringArray(targetFilePathList);
-        parser.createASTs(sourceFilePaths, null, new String[]{}, requestor, new NullProgressMonitor());
-
-        return fileASTList;
-    }
-
-    @Override
-    public abstract ASTParser createParser();
-
-    @SuppressWarnings("unchecked")
-    Map<String, String> getOptionMap() {
-        final Map<String, String> options = DefaultCodeFormatterConstants.getEclipseDefaultSettings();
-        options.put(JavaCore.COMPILER_COMPLIANCE, JavaCore.VERSION_10);
-        options.put(JavaCore.COMPILER_CODEGEN_TARGET_PLATFORM, JavaCore.VERSION_10);
-        options.put(JavaCore.COMPILER_SOURCE, JavaCore.VERSION_10);
-        return options;
-    }
-
-    List<Path> getTargetFiles(final Path targetFileRootPath, final String targetExtension) {
-        try {
-            return Files.walk(targetFileRootPath)
-                    .filter(path -> path.toString().endsWith(targetExtension))
-                    .collect(Collectors.toList());
-        } catch (IOException e) {
-            throw new RuntimeException(INVALID_PATH_EXCEPTION_MESSAGE);
-        }
-    }
-
-    String[] convertPathListToStringArray(final List<Path> pathList) {
-        return pathList.stream()
-                .map(Path::toString)
-                .toArray(String[]::new);
-    }
-}
