+++ src/main/java/com/github/t45k/abcd/ast/constructor/IASTConstructor.java
@@ -7,7 +7,7 @@ import java.nio.file.Path;
 import java.util.List;
 
 public interface IASTConstructor {
-    public List<FileAST> constructFileAST(final Path targetFilesRootPath);
+    List<FileAST> constructFileAST(final Path targetFilesRootPath);
 
-    public ASTParser createParser();
+    ASTParser createParser();
 }
