+++ src/main/java/com/github/t45k/abcd/ast/constructor/NonBindingASTConstructor.java
@@ -1,0 +1,18 @@
+package com.github.t45k.abcd.ast.constructor;
+
+import org.eclipse.jdt.core.dom.AST;
+import org.eclipse.jdt.core.dom.ASTParser;
+
+import java.util.Map;
+
+public class NonBindingASTConstructor extends ASTConstructor {
+
+    @Override
+    public ASTParser createParser() {
+        final ASTParser parser = ASTParser.newParser(AST.JLS10);
+        final Map<String, String> options = getOptionMap();
+        parser.setCompilerOptions(options);
+
+        return parser;
+    }
+}
